Something is working

+++ src/main/java/ksr/zad2/soft/functions/MembershipFunction.java
@@ -2,5 +2,5 @@ package ksr.zad2.soft.functions;
 
 public interface MembershipFunction<T> {
 
-    double calculate(T t);
+    float calculate(T t);
 }
